refactor: move useBalanceDetails to distinct file
