+++ src/components/Send.jsx
@@ -7,9 +7,10 @@ import PageTitle from './PageTitle'
 import ToggleSwitch from './ToggleSwitch'
 import Sprite from './Sprite'
 import Balance from './Balance'
-import { useReloadCurrentWalletInfo, useCurrentWallet, useBalanceDetails } from '../context/WalletContext'
+import { useReloadCurrentWalletInfo, useCurrentWallet } from '../context/WalletContext'
 import { useServiceInfo, useReloadServiceInfo } from '../context/ServiceInfoContext'
 import { useSettings } from '../context/SettingsContext'
+import { useBalanceDetails } from '../hooks/BalanceDetails'
 import * as Api from '../libs/JmWalletApi'
 import { routes } from '../constants/routes'
 import styles from './Send.module.css'
